# contracts/openapi/kelimio-api.yaml
openapi: 3.1.0
info:
  title: Kelimio API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Versioned contract for the Kelimio mobile and administrative clients. Server
    state is authoritative for identity, answers, score, energy, progress, and
    entitlements.
...
servers:
  - url: /
    description: Base URL supplied by the environment-specific client configuration.
...
security:
  - bearerAuth: []
paths:
...
  /v1/catalog/courses:
    get:
      operationId: listCatalogCourses
      tags: [Catalog]
      summary: List public, published courses
      parameters:
        - $ref: "#/components/parameters/ClientCapabilities"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: targetLanguage
          in: query
          schema:
            $ref: "#/components/schemas/LanguageTag"
        - name: supportLanguage
          in: query
          schema:
            $ref: "#/components/schemas/LanguageTag"
      responses:
        "200":
          description: A cursor page of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoursePage"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/courses/{courseId}:
    get:
      operationId: getCourse
      tags: [Catalog]
      summary: Return course details visible to the authenticated user
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/ClientCapabilities"
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
  /v1/courses/{courseId}/enrollments:
    post:
      operationId: enrollInCourse
      tags: [Enrollment]
      summary: Enroll the authenticated user in a free public course
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/ClientCapabilities"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnrollmentRequest"
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "200":
          description: Existing enrollment returned for an idempotent retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token obtained with Authorization Code and PKCE.
  headers:
...
    NoStore:
      description: Prevents sensitive answer feedback and owner-scoped import data from being cached.
      schema:
        type: string
        const: no-store
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
    Cursor:
      name: cursor
      in: query
      description: >-
        Opaque tamper-evident cursor bound to the authenticated owner, import,
        immutable preview/report identity, and position. It contains no
        workbook text or storage coordinates.
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable UUID generated once for the logical command.
      schema:
        type: string
        format: uuid
    ClientCapabilities:
      name: X-Kelimio-Client-Capabilities
      in: header
      required: false
      description: >-
        Comma-separated bounded capability tokens implemented by the client.
        Missing means the empty set; this is a compatibility signal, not authorization.
      schema:
        type: string
        minLength: 1
        maxLength: 1040
        pattern: "^[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9]*)+(?:, ?[a-z][a-z0-9]*(?:\\.[a-z][a-z0-9]*)+)*$"
  responses:
    Problem:
      description: RFC 9457 problem details
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Id:
      type: string
      format: uuid
...
    LanguageTag:
      type: string
      description: >-
        Canonically cased BCP 47 subset: lowercase primary language, optional
        title-case script, uppercase region, and lowercase variants. Extensions
        and private-use subtags are outside the initial API contract.
      minLength: 2
      maxLength: 35
      pattern: "^[a-z]{2,8}(?:-[A-Z][a-z]{3})?(?:-(?:[A-Z]{2}|[0-9]{3}))?(?:-(?:[a-z0-9]{5,8}|[0-9][a-z0-9]{3}))*$"
      examples: [tr, en, ar, fr, pt-BR, zh-Hant]
...
    CourseSummary:
      type: object
      additionalProperties: false
      required:
        [id, name, targetLanguage, supportLanguages, accessType, visibility, enrolled]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: [string, "null"]
          maxLength: 2000
        targetLanguage:
          $ref: "#/components/schemas/LanguageTag"
        supportLanguages:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/LanguageTag"
        accessType:
          type: string
          enum: [FREE, PAID]
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        enrolled:
          type: boolean
    CourseDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - targetLanguage
        - supportLanguages
        - accessType
        - visibility
        - enrolled
        - ownerDisplayName
        - releaseId
        - tests
      properties:
        id:
          $ref: "#/components/schemas/Id"
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: [string, "null"]
          maxLength: 2000
        targetLanguage:
          $ref: "#/components/schemas/LanguageTag"
        supportLanguages:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/LanguageTag"
        accessType:
          type: string
          enum: [FREE, PAID]
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        enrolled:
          type: boolean
        ownerDisplayName:
          type: string
          maxLength: 80
        releaseId:
          $ref: "#/components/schemas/Id"
        tests:
          type: array
          items:
            $ref: "#/components/schemas/TestSummary"
...
    TestSummary:
      type: object
      additionalProperties: false
      required: [id, revisionId, name, position, questionCount]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        revisionId:
          $ref: "#/components/schemas/Id"
        name:
          type: string
          maxLength: 160
        position:
          type: integer
          minimum: 1
        questionCount:
          type: integer
          minimum: 1
    CoursePage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CourseSummary"
        nextCursor:
          type: [string, "null"]
    CreateEnrollmentRequest:
      type: object
      additionalProperties: false
      required: [supportLanguage]
      properties:
        supportLanguage:
          $ref: "#/components/schemas/LanguageTag"
    EnrollmentResponse:
      type: object
      additionalProperties: false
      required: [id, courseId, supportLanguage, status, enrolledAt]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        courseId:
          $ref: "#/components/schemas/Id"
        supportLanguage:
          $ref: "#/components/schemas/LanguageTag"
        status:
          type: string
          enum: [ACTIVE]
        enrolledAt:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        requestId:
          type: string